moved common-utils
