+++ src/CMakeLists.txt
@@ -1,0 +1,3 @@
+if (BUILD_THERION OR BUILD_LOCH)
+    add_subdirectory(common-utils)
+endif()
