+++ src/common-utils/lxMath.cxx
